Extracting icon types
