+++ amundsen_application/static/js/components/common/SVGIcons/index.tsx
@@ -4,14 +4,3 @@
 export * from './AlertIcon';
 export * from './DownIcon';
 export * from './UpIcon';
-
-export enum IconSizes {
-  REGULAR = 24,
-  SMALL = 16,
-}
-
-export interface IconProps {
-  stroke?: string;
-  size?: number;
-  fill?: string;
-}
