+++ amundsen_application/static/js/components/common/SVGIcons/types.ts
@@ -1,0 +1,13 @@
+// Copyright Contributors to the Amundsen project.
+// SPDX-License-Identifier: Apache-2.0
+
+export enum IconSizes {
+  REGULAR = 24,
+  SMALL = 16,
+}
+
+export interface IconProps {
+  stroke?: string;
+  size?: number;
+  fill?: string;
+}
